refactor : #57 boardContent 컴포넌트 분리
