+++ project_frontend/src/components/index.js
@@ -1,16 +1,17 @@
+import Header from './header';
+import Dropdown from './dropdown';
+import GetlocationButton from './getLocationButton';
 import KakaoLoginButton from './login/kakaoLoginButton';
 import JoinComment from './login/joinComment';
 import EditorComponent from './write/editorComponent';
-import Header from './header';
 import WriteForm from './write/writeForm';
 import WriteWrapper from './write/writeWrapper';
-import GetlocationButton from './getLocationButton';
-import Dropdown from './dropdown';
 import { CustomSelectActivity } from './write/customSelect';
 import { InfiniteScroll } from './main/infiniteScroll';
 import { BoardItemsWrap } from './main/mainStyledComponents';
 import { Card } from './main/card';
 import { LinkTo, Profile } from './globalStyledComponents';
+import { BoardContent } from './boardDetails/boardContent';
 import {
 	BoardWrapper,
 	HeadingDiv,
@@ -19,7 +20,9 @@ import {
 	VerticalBar,
 	ContentDiv,
 	BoardInfo,
+	ReplyDiv,
 	ReplyInput,
+	ReplySubmitButton,
 } from './boardDetails/boardDetailsStyledComponents';
 
 export {
@@ -33,6 +36,7 @@ export {
 	CustomSelectActivity,
 	InfiniteScroll,
 	Card,
+	BoardContent,
 };
 export {
 	WriteWrapper,
@@ -46,5 +50,7 @@ export {
 	VerticalBar,
 	ContentDiv,
 	BoardInfo,
+	ReplyDiv,
 	ReplyInput,
+	ReplySubmitButton,
 };
